backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Shoaib",email = "shoaibulhaqq@outlook.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "uvicorn[standard] (>=0.23.2,<0.24.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-groq (>=0.3.6,<0.4.0)",
    "langchain-community[duckduckgo] (>=0.3.27,<0.4.0)",
    "duckduckgo-search (>=8.1.1,<9.0.0)",
    "elevenlabs (>=2.8.1,<3.0.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
